fixed the bug of saving product offering in user, collection and asking products collection
now, when the user send a product to firebase, the product along with the asking products, and all the corresponding images, will be saved in both user object and product offerings collection and asking products collection.  I save this info duplicately is because it will reduce the number of queries required to search for the products info.