done processing a bid.
the bid form screen creates the bid when the send button is clicked.  if the info is valid, the form create the bid, then it proceed to call the process bidding function in the bid view model.  FirebaseClient will then process the bid and save the bid in the product bidding and send to firestore.