done refactoring accept bid's interface
building accept bid function.  by accepting the bid, the transaction will be closed before elapse time.  The app will write to firestore's acceptBids collection, which triggers the cloud function to process the accetance.  I am writing the cloud function to delete the product offering from collection and users' objects.  It also added an accepted bid object to users' objects.  so, the app can parse these objects and show in transactions in the app.