refactoring the product classes dateCreated field
I want to display a standard date and time in the products offered, and in the bids.  I decided to use "America/Toronto" time zone as the server time.  I store the time string in UTC time format, in various classes.  the app will parse the string to a date time object and extract the time info.  The app does various computations and ranking then.