done the delete product function.
users (the seller of the product), now can delete the product he is selling.  It will be removed from the productsOffering collection and the user object in firestore.

But I still need to remove the product in the local database too.  I will do it later.

I'm refactoring the interfaces.  there is also a bug that the screen composable raised an exception of reading a product.