implementing the delete image function in user(owner)'s product offereing
already implemented delete image when the images are not yet saved in sell screen, asking product screen and bid form screen.  implementing the delete image function for the product offering which is already saved in firestore.  we need to delete the image url from the product in local database.  We also need to send a request to update the product object in firestore.  I am modifying the Product Image class to add the download url in the class.  so, we can identify the image and remove it.  however, I also like to refactor the app to save the images in local storage, and save the local url in an image class.  So, if we have the image in the device, we don't need to send a request to cloud storage to download it.