refactoring the image loading mechanisms in the database manager
the database manager is responsible for loading images from local database, or from the cloud.  so, the other screens of the app can get the images to display.

the image loading mechanism is to first try to load the image from local database, using the url from the product offering, images field.  if it returns null, it will retrieve the image from Cloud Storage.  and save the image in the local database.