moved the loading of accept bids from bid details vm to database mgr
further refactoring the app to do all the retrievals of data from the local database manager.  Moved the loading of bids from bid details vm to the mgr.  trying to separate the accepted bids (the seller user who accepts the bid from the buyer) and the bid accepted (the buyer user who makes the bid, and the bid got accepted).  need to display them separately.