fixed images saving and retrievals
now, the app first try to retrieve the images from local storage.  If it can't find it, it will download from the cloud storage and then save into local storage.

the interface uses loadedOrPlaceholder composable and view model to load the image from local storage, or download from the cloud.