finished messaging functions.
now, buyer and seller can message each other after the acceptance of a bid.  The messages are sent to Firestore.  A cloud function writes the messages to the target user object.

working on closing transaction.  either buyer or seller can initiate the request of closing the transaction.  the other party needs to confirm it.  Then, the transaction is closed.  The product will be removed from firestore's products list.