handle expired products
for the products that passed expiry, I set it's status to -1 and -2.  -1 is expired less than 20 days, -2 is more than 20, that will be deleted.

the products which are expired should not be shown in user's products and products available for bids.  The products which are successfully exchanged, status are set to 5, or 6 for transaction closed.  They appear in accepted bids.

fixed small bugs in send message screen.